# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/api/space/createQuestion": {
            "post": {
                "tags": [
                    "Question"
                ],
                "summary": "user create question",
                "description": "A user create question",
                "operationId": "82d22483199fc6f2189b2b15c27b8a87",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "question_desc"
                                ],
                                "properties": {
                                    "question_desc": {
                                        "type": "required|string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "api/user/question/{user_id}": {
            "get": {
                "tags": [
                    "Question"
                ],
                "summary": "Get all question from users",
                "description": "Get all questions from users",
                "operationId": "cb12a3f7aeb559a776c6dc5dabf3496f",
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProjectResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unable to get question"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
    }
}